fix overflow moving from usize to u32 by increasing to u64

+++ src/day06.rs
@@ -6,33 +6,33 @@ use nom::IResult;
 
 #[derive(Debug, PartialEq)]
 struct TimeAndDistance {
-    time: u32,
-    distance: u32,
+    time: u64,
+    distance: u64,
 }
 
 impl TimeAndDistance {
-    fn distance_travelled(&self, held: u32) -> u32 {
+    fn distance_travelled(&self, held: u64) -> u64 {
         self.time.saturating_sub(held).saturating_mul(held)
     }
 
-    fn winning_possibilities(&self) -> u32 {
+    fn winning_possibilities(&self) -> u64 {
         (1..(self.time - 1))
             .map(|t| self.distance_travelled(t))
             .skip_while(|d| *d <= self.distance)
             .take_while(|d| *d > self.distance)
-            .count() as u32
+            .count() as u64
     }
 }
 
-fn parse_numbers(input: &str) -> IResult<&str, Vec<u32>> {
-    many1(preceded(take_while(char::is_whitespace), complete::u32))(input)
+fn parse_numbers(input: &str) -> IResult<&str, Vec<u64>> {
+    many1(preceded(take_while(char::is_whitespace), complete::u64))(input)
 }
 
-fn parse_time(input: &str) -> IResult<&str, Vec<u32>> {
+fn parse_time(input: &str) -> IResult<&str, Vec<u64>> {
     preceded(tag("Time:"), parse_numbers)(input)
 }
 
-fn parse_distance(input: &str) -> IResult<&str, Vec<u32>> {
+fn parse_distance(input: &str) -> IResult<&str, Vec<u64>> {
     preceded(tag("Distance:"), parse_numbers)(input)
 }
 
@@ -47,16 +47,16 @@ fn input_into_time_and_distance(input: &str) -> Vec<TimeAndDistance> {
         .collect()
 }
 
-fn parse_numbers2(input: &str) -> IResult<&str, u32> {
+fn parse_numbers2(input: &str) -> IResult<&str, u64> {
     let (remainder, strings) = many1(preceded(take_while(char::is_whitespace), digit1))(input)?;
     Ok((remainder, strings.join("").parse().unwrap()))
 }
 
-fn parse_time2(input: &str) -> IResult<&str, u32> {
+fn parse_time2(input: &str) -> IResult<&str, u64> {
     preceded(tag("Time:"), parse_numbers2)(input)
 }
 
-fn parse_distance2(input: &str) -> IResult<&str, u32> {
+fn parse_distance2(input: &str) -> IResult<&str, u64> {
     preceded(tag("Distance:"), parse_numbers2)(input)
 }
 
@@ -70,7 +70,7 @@ pub fn part1(input: &str) -> String {
     input_into_time_and_distance(input)
         .into_iter()
         .map(|dt| dt.winning_possibilities())
-        .product::<u32>()
+        .product::<u64>()
         .to_string()
 }
